+++ packages/agent-browser-extension/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @agent-native/agent-browser-extension
 
+## 0.1.269
+
+### Patch Changes
+
+- Updated dependencies [5a045bf]
+- Updated dependencies
+  - @agent-native/core@0.175.1
+
+## 0.1.268
+
+### Patch Changes
+
+- Updated dependencies [73ff8c5]
+- Updated dependencies [da836e2]
+- Updated dependencies
+- Updated dependencies [de5ba2d]
+- Updated dependencies [cf473dc]
+- Updated dependencies [6c71a21]
+  - @agent-native/core@0.175.0
+
 ## 0.1.267
 
 ### Patch Changes